update test_api_db contract and api_tests/db_tests to use the updated contract; also improve api_tests/action_receipt_tests
