+++ unittests/test-contracts/test_api_db/CMakeLists.txt
@@ -1,5 +1,6 @@
 if( ${eosio.cdt_FOUND} )
-  add_executable( test_api_db test_api_db.cpp )
+   add_contract( test_api_db test_api_db test_api_db.cpp )
 else()
-  configure_file( ${CMAKE_CURRENT_SOURCE_DIR}/test_api_db.wasm ${CMAKE_CURRENT_BINARY_DIR}/test_api_db.wasm COPYONLY )
+   configure_file( ${CMAKE_CURRENT_SOURCE_DIR}/test_api_db.wasm ${CMAKE_CURRENT_BINARY_DIR}/test_api_db.wasm COPYONLY )
+   configure_file( ${CMAKE_CURRENT_SOURCE_DIR}/test_api_db.abi  ${CMAKE_CURRENT_BINARY_DIR}/test_api_db.abi  COPYONLY )
 endif()
